Разбивает код проекта на модули.
